+++ src/model/factory.ts
@@ -9,6 +9,7 @@ import { Histogram } from './histogram';
 import { LineTrace } from './line';
 import { ScatterTrace } from './scatter';
 import { SegmentedTrace } from './segmented';
+import { SmoothTrace } from './smooth';
 
 export abstract class TraceFactory {
   public static create(layer: MaidrLayer): Trace {
@@ -33,6 +34,9 @@ export abstract class TraceFactory {
 
       case TraceType.SCATTER:
         return new ScatterTrace(layer);
+
+      case TraceType.SMOOTH:
+        return new SmoothTrace(layer);
 
       case TraceType.DODGED:
       case TraceType.NORMALIZED: